Rename presence source files
