+++ src/panel_app_main.cpp
@@ -1,7 +1,7 @@
+#include "bus_monitor.hpp"
 #include "button_handler.hpp"
 #include "const.hpp"
 #include "panel_app.hpp"
-#include "presence.hpp"
 #include "utils.hpp"
 
 #include <exception>
